fixed hopper to use only 2 motor sections - The Cult of Mason
