+++ src/main/java/org/team3128/compbot/subsystems/Hopper.java
@@ -20,10 +20,12 @@ public class Hopper extends Threaded {
 
     public enum HopperState {
         POS_0(new boolean[] { false, false, false, false, false }),
-        POS_1(new boolean[] { false, false, false, false, true }),
-        POS_2(new boolean[] { false, false, false, true, true }),
-        POS_3(new boolean[] { false, false, true, true, true }), POS_4(new boolean[] { false, true, true, true, true }),
-        POS_5(new boolean[] { true, true, true, true, true });
+        POS_1(new boolean[] { true, false, false, false, false }),
+        POS_2(new boolean[] { true, true, false, false, false }),
+        POS_3(new boolean[] { true, true, false, true, false }),
+        POS_4(new boolean[] { true, true, true, true, false }),
+        POS_5(new boolean[] { true, true, true, true, true }),
+        POS_6(new boolean[] { true, true, true, false, false });
 
         public boolean[] hopperState;
 
@@ -32,9 +34,9 @@ public class Hopper extends Threaded {
         }
     }
 
-    public LazyCANSparkMax INTAKE_MOTOR, HOPPER_FEEDER_MOTOR, CORNER_MOTOR, SHOOTER_FEEDER_MOTOR, GATEKEEPER_MOTOR;
-    public CANEncoder HOPPER_FEEDER_ENCODER, CORNER_ENCODER, SHOOTER_FEEDER_ENCODER;
-    public DigitalInput SENSOR_0, SENSOR_1, SENSOR_2, SENSOR_3, SENSOR_4;
+    public LazyCANSparkMax INTAKE_MOTOR, HOPPER_FEEDER_MOTOR, CORNER_MOTOR, GATEKEEPER_MOTOR;
+    public CANEncoder HOPPER_FEEDER_ENCODER, CORNER_ENCODER;
+    public DigitalInput SENSOR_0, SENSOR_1, SENSOR_2;
 
     boolean[] ballArray = { false, false, false, false, false };
     private static final Hopper instance = new Hopper();
@@ -43,7 +45,7 @@ public class Hopper extends Threaded {
     private boolean inPlaceExit = false;
     private boolean isMoving;
 
-    private DigitalInput[] sensorPositions = { SENSOR_0, SENSOR_1, SENSOR_2, SENSOR_3, SENSOR_4 }; // top to bottom
+    private DigitalInput[] sensorPositions = { SENSOR_0, SENSOR_1, SENSOR_2 }; // top to bottom //0 = 1.5, 1 = 3, 2 = 4 mathhh
 
     public static Hopper getInstance() {
         return instance;
@@ -76,17 +78,20 @@ public class Hopper extends Threaded {
     @Override
     public void update() {
         if (!isMoving) {
-            checkSensors();
+            checkSensors(); //TODO: do we need this?
         }
         // countBalls(SENSOR_0, SENSOR_4);
     }
 
-    private void checkSensors() {
+    /*private void checkSensors() {
         ballArray[0] = SENSOR_0.get();
         ballArray[1] = SENSOR_1.get();
         ballArray[2] = SENSOR_2.get();
         ballArray[3] = SENSOR_3.get();
         ballArray[4] = SENSOR_4.get();
+    }*/
+    private void checkSensors() {
+        
     }
 
     private void configMotors() {
@@ -94,23 +99,18 @@ public class Hopper extends Threaded {
         HOPPER_FEEDER_MOTOR = new LazyCANSparkMax(Constants.HopperConstants.HOPPER_FEEDER_MOTOR_ID,
                 MotorType.kBrushless);
         CORNER_MOTOR = new LazyCANSparkMax(Constants.HopperConstants.CORNER_MOTOR_ID, MotorType.kBrushless);
-        SHOOTER_FEEDER_MOTOR = new LazyCANSparkMax(Constants.HopperConstants.SHOOTER_FEEDER_MOTOR_ID,
-                MotorType.kBrushless);
         GATEKEEPER_MOTOR = new LazyCANSparkMax(Constants.HopperConstants.GATEKEEPER_MOTOR_ID, MotorType.kBrushless);
     }
 
     private void configEncoders() {
         HOPPER_FEEDER_ENCODER = HOPPER_FEEDER_MOTOR.getEncoder();
         CORNER_ENCODER = CORNER_MOTOR.getEncoder();
-        SHOOTER_FEEDER_ENCODER = SHOOTER_FEEDER_MOTOR.getEncoder();
     }
 
     private void configSensors() {
         SENSOR_0 = new DigitalInput(Constants.HopperConstants.SENSOR_0_ID);
         SENSOR_1 = new DigitalInput(Constants.HopperConstants.SENSOR_1_ID);
         SENSOR_2 = new DigitalInput(Constants.HopperConstants.SENSOR_2_ID);
-        SENSOR_3 = new DigitalInput(Constants.HopperConstants.SENSOR_3_ID);
-        SENSOR_4 = new DigitalInput(Constants.HopperConstants.SENSOR_4_ID);
     }
 
     public void setBallOrder(boolean[] CurrBallArray) {
@@ -148,15 +148,33 @@ public class Hopper extends Threaded {
 
     private boolean[] shift(boolean[] in_array) {
         boolean[] out_array = new boolean[in_array.length];
-        for (int i = 0; i < in_array.length - 2; i++) {
+        for (int i = 0; i < in_array.length - 1; i++) {
             out_array[i] = in_array[i + 1];
         }
         out_array[in_array.length - 1] = false;
         return out_array;
     }
 
+    private boolean[] addBall(boolean[] in_array) {
+        boolean[] out_array = new boolean[in_array.length];
+        for (int i = 0; i < in_array.length; i++) {
+            boolean added = false;
+            if(in_array[i]) {
+                out_array[i] = in_array[i];
+            } else if (!added) {
+                out_array[i] = true;
+                added = true;
+            }
+        }
+        return out_array;
+    }
+
     public boolean[] getBallArray() {
         return ballArray;
+    }
+
+    public void updateBallArray(boolean[] in_array) {
+        this.ballArray = in_array;
     }
 
     public void setIsMoving(boolean isMoving) {
@@ -173,5 +191,12 @@ public class Hopper extends Threaded {
             tempCount++;
         }
         return record_holder_index + 1;
+    }
+
+    public void setMotorPowers(double p_Gatekeeer, double p_Corner, double p_HopperFeeder) {
+        GATEKEEPER_MOTOR.set(p_Gatekeeer);
+        //SHOOTER_FEEDER_MOTOR.set(p_ShooterFeeder);
+        CORNER_MOTOR.set(p_Corner);
+        HOPPER_FEEDER_MOTOR.set(p_HopperFeeder);
     }
 }
